[FrontEnd] Add api interceptors
