--- src/data/API.ts
@@ -1,17 +1,0 @@
-import axios from 'axios';
-
-const api = axios.create({
-  baseURL: process.env.NEXT_PUBLIC_API_URL || 'http://localhost:3333',
-  timeout: 30000,
-  headers: {
-    Accept: 'application/json',
-    'Content-Type': 'application/json'
-  }
-});
-
-api.interceptors.response.use(
-  response => response,
-  error => error
-);
-
-export default api;
