+++ src/data/client/room.client.ts
@@ -1,7 +1,7 @@
 import {API_ENDPOINTS} from '@/configs/endpoint.config';
 import {IBaseResponse} from '@/types';
 
-import api from '../api';
+import api from '../http';
 import {IStoryResponse} from './story.client';
 import {IUserRoomResponse} from './userRoom.client';
